feat: Implementation of a caching system
- Increased the limit to display 50 posts to observe the caching mechanism in action.
- Introduced a dedicated store for the caching mechanism.
- The idea is to initiate a counter upon the loading of the index page,
and from there, a function that will invalidate the cache will be periodically launched.
- So, once the caching begins, every attempt to fetch more posts will display locally stored posts.
When the cache is invalidated after 30 seconds in our example, it becomes possible to fetch new posts from the API.
- In practice, following the login process, 10 posts are initially loaded, and you have the option to load an additional 10. Subsequent attempts to load more posts will display cached content only.
To trigger an API call, wait for around 30 seconds (tip: watch for this message in the console: '-- 909 -> Cache has been invalidated'). After this, if you attempt to load more posts, a new API call will be made, refreshing the cache once again.
When the cache is utilized, even after refreshing the page, offline data should be retrieved, even in the event of an unreachable API.
- Don't worry about all the messages on the console; I haven't used a logger for this particular implementation. In a real application, I would typically use one. However, for development purposes only. Forgive me for this practice

Management of the local database.

- As mentioned in the previous commit, I switched from PouchdbORM to the 'Localbase' module
(available on GitHub: /dannyconnell/localbase).
- I find the module's API more stable and easier to access.
- Moreover, it uses the same API as Firebase, and personally, I prefer the collection/documents structure.

Management of post display:

- This part is highly improvable.
- A composable 'useCollection' was created to handle all collection-related tasks.
- It should be easily reusable if we want to manage something other than posts.

Todo: There are still many things to improve, but the current priority is to load all posts on startup.
If time permits, I will refactor the application as some parts have become a bit complex.