Feat: Continuous loading of posts and caching mechanism.
Creation of a new composable "useProgression" to encapsulate the progression logic
and removal of the file responsible for synchronization.

Post Management:
Caching:
Consolidated all cache-related functionality in a dedicated "cache" store.
- I divided the logic of the "loadPosts" method into two parts:
- Handling new posts from the API when the cache is invalidated or when the user logs out.
- Managing cached posts.

- Continued using the "Localbase" module to write posts to the "livenews" IndexedDB database (posts collection).
- During the initial load (after the progress page, for example), if there are no cached posts, we make an API call and save all posts locally (default 30 returned by the API).
- Once a local copy is saved, we attempt to load future posts in batches of 10.

Note: I say "attempt" because there are likely still improvements to be made on this front. But, in general, it works.
- To save posts in batches of 10, I use an "offset" to calculate the necessary range to display the correct posts.
- On the index page, a function runs continuously to invalidate the cache every x time. This x can be defined as a parameter of the "scheduleInvalidation" method.
Reset:

When the user logs out, I reset the posts.

Note: The application still has bugs, especially in the synchronization of posts to display under certain conditions.

Bugfix: Finally, continuous loading of posts is possible. There was a bug with my use of the Quasar infinite scroll component and Localbase.
I had to track the result coming from Localbase in a computed property.
This resolved the bug.

IMPORTANT: To fetch new posts from the API again, follow these steps:

Either clear the value of "cacheValid" in local storage.
Wait for the specified time in the "scheduleInvalidation" method in indexPage.vue (set to a shorter value).
Log out.

Acknowledgment: Thank you for this very enlightening test. Best wishes to those reading this.